+++ Realm/ObjectServerTests/SwiftObjectServerTests.swift
@@ -86,6 +86,72 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
             XCTFail("Got an error: \(error) (process: \(isParent ? "parent" : "child"))")
         }
     }
+
+    // MARK: Permissions
+
+    func testPermissionChange() {
+        do {
+            let userA = try synchronouslyLogInUser(for: basicCredential(register: isParent, usernameSuffix: "_A"), server: authURL)
+            let userB = try synchronouslyLogInUser(for: basicCredential(register: isParent, usernameSuffix: "_B"), server: authURL)
+            _ = try synchronouslyOpenRealm(url: realmURL, user: userA)
+
+            let adminPermissions = [
+                [true, true, true],
+                [false, true, true],
+                [true, false, true],
+                [false, false, true]
+            ]
+            let readWritePermissions = [[true, true, false]]
+            let readOnlyPermissions = [[true, false, false]]
+            let noAccessPermissions: [[Bool?]] = [
+                [false, false, false],
+                [nil, nil, nil]
+            ]
+            let permissions = [adminPermissions, readWritePermissions, readOnlyPermissions, noAccessPermissions]
+            let statusMessages = [
+                "administrative access",
+                "read-write access",
+                "read-only access",
+                "no access"
+            ]
+
+            for (accessPermissions, statusMessage) in zip(permissions, statusMessages) {
+                for permissions in accessPermissions {
+                    let permissionChange = SyncPermissionChange(
+                        realmURL: realmURL.absoluteString,
+                        userID: userB.identity!,
+                        mayRead: permissions[0],
+                        mayWrite: permissions[1],
+                        mayManage: permissions[2]
+                    )
+                    try verifyChangePermission(change: permissionChange, statusMessage: statusMessage, owner: userA)
+                }
+            }
+        } catch {
+            XCTFail("Got an error: \(error) (process: \(isParent ? "parent" : "child"))")
+        }
+    }
+
+    func verifyChangePermission(change: SyncPermissionChange, statusMessage: String, owner: SyncUser) throws {
+        let managementRealm = try owner.managementRealm()
+
+        let exp = expectation(description: "A new permission will be granted by the server")
+        let token = managementRealm.objects(SyncPermissionChange.self).filter("id = %@", change.id).addNotificationBlock { changes in
+            if case .update(let change, _, _, _) = changes, let statusCode = change[0].statusCode.value {
+                XCTAssertEqual(statusCode, 0)
+                XCTAssertEqual(change[0].status, .success)
+                XCTAssertNotNil(change[0].statusMessage?.range(of: statusMessage))
+                exp.fulfill()
+            }
+        }
+
+        try managementRealm.write {
+            managementRealm.add(change)
+        }
+
+        waitForExpectations(timeout: 2)
+        token.stop()
+    }
 }
 #else
 class SwiftObjectServerTests: SwiftSyncTestCase {
@@ -153,6 +219,72 @@ class SwiftObjectServerTests: SwiftSyncTestCase {
         } catch {
             XCTFail("Got an error: \(error) (process: \(isParent ? "parent" : "child"))")
         }
+    }
+
+    // MARK: Permissions
+
+    func testPermissionChange() {
+        do {
+            let userA = try synchronouslyLogInUser(for: basicCredential(register: isParent, usernameSuffix: "_A"), server: authURL)
+            let userB = try synchronouslyLogInUser(for: basicCredential(register: isParent, usernameSuffix: "_B"), server: authURL)
+            _ = try synchronouslyOpenRealm(url: realmURL, user: userA)
+
+            let adminPermissions: [[Bool?]] = [
+                [true, true, true],
+                [false, true, true],
+                [true, false, true],
+                [false, false, true]
+            ]
+            let readWritePermissions: [[Bool?]] = [[true, true, false]]
+            let readOnlyPermissions: [[Bool?]] = [[true, false, false]]
+            let noAccessPermissions: [[Bool?]] = [
+                [false, false, false],
+                [nil, nil, nil]
+            ]
+            let permissions: [[[Bool?]]] = [adminPermissions, readWritePermissions, readOnlyPermissions, noAccessPermissions]
+            let statusMessages = [
+                "administrative access",
+                "read-write access",
+                "read-only access",
+                "no access"
+            ]
+
+            for (accessPermissions, statusMessage) in zip(permissions, statusMessages) {
+                for permissions in accessPermissions {
+                    let permissionChange = SyncPermissionChange(
+                        realmURL: realmURL.absoluteString,
+                        userID: userB.identity!,
+                        mayRead: permissions[0],
+                        mayWrite: permissions[1],
+                        mayManage: permissions[2]
+                    )
+                    try verifyChangePermission(change: permissionChange, statusMessage: statusMessage, owner: userA)
+                }
+            }
+        } catch {
+            XCTFail("Got an error: \(error) (process: \(isParent ? "parent" : "child"))")
+        }
+    }
+
+    func verifyChangePermission(change change: SyncPermissionChange, statusMessage: String, owner: SyncUser) throws {
+        let managementRealm = try owner.managementRealm()
+
+        let exp = expectationWithDescription("A new permission will be granted by the server")
+        let token = managementRealm.objects(SyncPermissionChange.self).filter("id = %@", change.id).addNotificationBlock { changes in
+            if case .Update(let change, _, _, _) = changes, let statusCode = change[0].statusCode.value {
+                XCTAssertEqual(statusCode, 0)
+                XCTAssertEqual(change[0].status, SyncManagementObjectStatus.Success)
+                XCTAssertNotNil(change[0].statusMessage?.rangeOfString(statusMessage))
+                exp.fulfill()
+            }
+        }
+
+        try managementRealm.write {
+            managementRealm.add(change)
+        }
+
+        waitForExpectationsWithTimeout(2, handler: nil)
+        token.stop()
     }
 }
 #endif