+++ Realm/RLMSyncUser.mm
@@ -157,6 +157,10 @@ using namespace realm;
     }
 }
 
+- (RLMRealm *)managementRealmWithError:(NSError **)error {
+    return [RLMRealm realmWithConfiguration:[RLMRealmConfiguration managementConfigurationForUser:self] error:error];
+}
+
 #pragma mark - Private API
 
 - (void)_unregisterRefreshHandleForURLPath:(NSString *)path {